pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "render-kit"
version = "1.0.0"
description = "Generate product photos and social media images from the command line."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "NOUMENON", email = "hello@noumenon.ai"}]
keywords = ["product-photography", "social-media", "image-generation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
ai = ["replicate>=0.25.0"]
voice = ["edge-tts>=6.1.0"]
video = []

[project.scripts]
render-kit = "render_kit.cli:app"

[project.urls]
Homepage = "https://github.com/noumenon-ai/render-kit"
Repository = "https://github.com/noumenon-ai/render-kit"

[tool.setuptools.packages.find]
where = ["src"]
